refactor: consolidate page link handling and remove unused components
- Deleted the `page_links.py` module, consolidating its functionality into `probe_links.py` for better organization.
- Updated `probe_links.py` to include necessary classes and functions previously in `page_links.py`, ensuring all page link handling is centralized.
- Removed obsolete contract tests related to page link normalization, streamlining the test suite.
- Enhanced documentation in `probe_links.py` to clarify the handling of internal page hyperlinks.
