+++ apps/worker/tests/contract/test_probe_outline_links_contract.py
@@ -1,4 +1,4 @@
-"""Contract tests for probe.outline tree prune and probe.links noise markers."""
+"""Contract tests for probe.outline prune, probe.links noise, and dest page normalize."""
 
 from __future__ import annotations
 
@@ -11,8 +11,11 @@ os.environ.setdefault("S3_ACCESS_KEY_ID", "test")
 os.environ.setdefault("S3_SECRET_ACCESS_KEY", "test")
 os.environ.setdefault("S3_TEMP_PATH", "/tmp")
 
-from app.services.document_agent.structure.page_links import PageLink
-from app.services.document_agent.tools.probe_links import annotate_link_noise
+from app.services.document_agent.tools.probe_links import (
+    PageLink,
+    _link_dest_physical_page,
+    annotate_link_noise,
+)
 from app.services.document_agent.tools.probe_outline import build_outline_forest
 
 
@@ -47,7 +50,13 @@ def test_outline_drops_no_page_leaf_and_empty_subtree() -> None:
 
 def test_link_noise_marks_page_number_header_and_repeated_dest() -> None:
     links = [
-        PageLink(source_page=2, dest_physical_page=10, anchor_text="12", from_y0=5, page_height=100),
+        PageLink(
+            source_page=2,
+            dest_physical_page=10,
+            anchor_text="12",
+            from_y0=5,
+            page_height=100,
+        ),
         PageLink(
             source_page=2,
             dest_physical_page=10,
@@ -77,3 +86,11 @@ def test_link_noise_marks_page_number_header_and_repeated_dest() -> None:
     assert "repeated_dest" in by_anchor["12"]["noise"]
     assert "repeated_dest" in by_anchor["Intro"]["noise"]
     assert by_anchor["Normal title"]["noise"] == []
+
+
+def test_link_dest_physical_page_by_type() -> None:
+    """``int`` is 0-based; digit ``str`` is already 1-based."""
+    assert _link_dest_physical_page(6) == 7
+    assert _link_dest_physical_page(0) == 1
+    assert _link_dest_physical_page("6") == 6
+    assert _link_dest_physical_page("7") == 7
